change file structure re #1
