+++ .gitignore
@@ -1,0 +1,3 @@
+config.json
+.vscode
+node_modules
