+++ src/assets/images/icons/icon-128x128.png
