+++ src/assets/images/icons/icon-32x32.png
