+++ src/assets/images/scattered-showers.png
