+++ src/assets/images/thunderstorm.png
