+++ src/assets/images/wind.png
